refactor external-resizer to use csi-lib-utils/rpc
